+++ src/bgmCompare/bgmCompare_logp_and_grad.h
@@ -4,29 +4,6 @@
 
 
 
-double log_pseudoposterior(
-    const arma::mat& main_effects,
-    const arma::mat& pairwise_effects,
-    const arma::imat& main_effect_indices,
-    const arma::imat& pairwise_effect_indices,
-    const arma::mat& projection,
-    const arma::imat& observations,
-    const arma::imat& group_indices,
-    const arma::ivec& num_categories,
-    const std::vector<arma::imat>& counts_per_category,
-    const std::vector<arma::imat>& blume_capel_stats,
-    const std::vector<arma::mat>& pairwise_stats,
-    const int num_groups,
-    const arma::imat& inclusion_indicator,
-    const arma::uvec& is_ordinal_variable,
-    const arma::ivec& baseline_category,
-    const double main_alpha,
-    const double main_beta,
-    const double interaction_scale,
-    const arma::mat& pairwise_scaling_factors,
-    const double difference_scale
-);
-
 arma::uword total_length(
     const int num_variables,
     const arma::imat& main_effect_indices,
@@ -55,13 +32,36 @@ arma::vec gradient_observed_active(
   const arma::imat pair_index
 );
 
+double log_pseudoposterior(
+    const arma::mat& main_effects,
+    const arma::mat& pairwise_effects,
+    const arma::imat& main_effect_indices,
+    const arma::imat& pairwise_effect_indices,
+    const arma::mat& projection,
+    const arma::mat& observations_double,
+    const arma::imat& group_indices,
+    const arma::ivec& num_categories,
+    const std::vector<arma::imat>& counts_per_category_group,
+    const std::vector<arma::imat>& blume_capel_stats_group,
+    const std::vector<arma::mat>&  pairwise_stats_group,
+    const int num_groups,
+    const arma::imat& inclusion_indicator,
+    const arma::uvec& is_ordinal_variable,
+    const arma::ivec& baseline_category,
+    const double main_alpha,
+    const double main_beta,
+    const double interaction_scale,
+    const arma::mat& pairwise_scaling_factors,
+    const double difference_scale
+);
+
 arma::vec gradient(
     const arma::mat& main_effects,
     const arma::mat& pairwise_effects,
     const arma::imat& main_effect_indices,
     const arma::imat& pairwise_effect_indices,
     const arma::mat& projection,
-    const arma::imat& observations,
+    const arma::mat& observations_double,
     const arma::imat& group_indices,
     const arma::ivec& num_categories,
     const std::vector<arma::imat>& counts_per_category_group,